+++ public/components/incontext_insight/generate_popover_body.tsx
@@ -7,10 +7,10 @@ import React, { useState } from 'react';
 import { i18n } from '@osd/i18n';
 import { v4 as uuidv4 } from 'uuid';
 import {
-  EuiBadge,
-  EuiButtonEmpty,
   EuiFlexGroup,
   EuiFlexItem,
+  EuiIcon,
+  EuiIconTip,
   EuiLoadingContent,
   EuiMarkdownFormat,
   EuiPanel,
@@ -18,6 +18,7 @@ import {
   EuiPopoverTitle,
   EuiSpacer,
   EuiText,
+  EuiTitle,
 } from '@elastic/eui';
 import { useEffectOnce } from 'react-use';
 import { MessageActions } from '../../tabs/chat/messages/message_action';
@@ -62,9 +63,19 @@ export const GeneratePopoverBody: React.FC<{
 
   const onGenerateSummary = (summarizationQuestion: string) => {
     const summarize = async () => {
-      const contextObj = incontextInsight.contextProvider
-        ? await incontextInsight.contextProvider()
-        : undefined;
+      let contextObj;
+      try {
+        contextObj = (await incontextInsight.contextProvider?.()) ?? undefined;
+      } catch (e) {
+        console.error('Error executing contextProvider:', e);
+        toasts.addDanger(
+          i18n.translate('assistantDashboards.incontextInsight.generateSummaryError', {
+            defaultMessage: 'Generate summary error',
+          })
+        );
+        closePopover();
+        return;
+      }
       const contextContent = contextObj?.context || '';
       let summaryType: string;
       const endIndex = incontextInsight.key.indexOf('_', 0);
@@ -92,12 +103,12 @@ export const GeneratePopoverBody: React.FC<{
         .then((response) => {
           const summaryContent = response.summary;
           setSummary(summaryContent);
-          const insightAgentIdExists = !!response.insightAgentId;
+          const insightAgentIdExists = insightType !== undefined && response.insightAgentIdExists;
           setInsightAvailable(insightAgentIdExists);
           if (insightAgentIdExists) {
             onGenerateInsightBasedOnSummary(
-              response.insightAgentId,
               summaryType,
+              insightType,
               summaryContent,
               contextContent,
               `Please provide your insight on this ${summaryType}.`
@@ -119,7 +130,7 @@ export const GeneratePopoverBody: React.FC<{
   };
 
   const onGenerateInsightBasedOnSummary = (
-    insightAgentId: string,
+    summaryType: string,
     insightType: string,
     summaryContent: string,
     context: string,
@@ -129,7 +140,7 @@ export const GeneratePopoverBody: React.FC<{
       httpSetup
         ?.post(SUMMARY_ASSISTANT_API.INSIGHT, {
           body: JSON.stringify({
-            insightAgentId,
+            summaryType,
             insightType,
             summary: summaryContent,
             context,
@@ -173,44 +184,34 @@ export const GeneratePopoverBody: React.FC<{
   const renderInnerTitle = () => {
     return (
       <EuiPopoverTitle className="incontextInsightGeneratePopoverTitle" paddingSize="l">
-        {showInsight ? (
-          <EuiFlexGroup gutterSize="none">
-            <EuiFlexItem grow={false}>
-              <EuiButtonEmpty
+        <EuiFlexGroup gutterSize="xs" alignItems="center">
+          <EuiFlexItem grow={false}>
+            {showInsight ? (
+              <EuiIcon
                 aria-label="back-to-summary"
-                flush="left"
-                size="xs"
+                size="m"
                 onClick={() => {
                   setShowInsight(false);
                 }}
-                iconType="arrowLeft"
-                iconSide={'left'}
+                type="arrowLeft"
                 color={'text'}
-              >
-                {i18n.translate('assistantDashboards.incontextInsight.InsightWithRAG', {
-                  defaultMessage: 'Insight With RAG',
-                })}
-              </EuiButtonEmpty>
-            </EuiFlexItem>
-          </EuiFlexGroup>
-        ) : (
-          <EuiFlexGroup gutterSize="none">
-            <EuiFlexItem>
-              <div>
-                <EuiBadge
-                  aria-label="alert-assistant"
-                  color="hollow"
-                  iconType={shiny_sparkle}
-                  iconSide="left"
-                >
+              />
+            ) : (
+              <EuiIcon aria-label="alert-assistant" color="hollow" size="l" type={shiny_sparkle} />
+            )}
+          </EuiFlexItem>
+          <EuiFlexItem>
+            <EuiText>
+              <EuiTitle size="xxs">
+                <h6>
                   {i18n.translate('assistantDashboards.incontextInsight.Summary', {
-                    defaultMessage: 'Summary',
+                    defaultMessage: showInsight ? 'Insight With RAG' : 'Summary',
                   })}
-                </EuiBadge>
-              </div>
-            </EuiFlexItem>
-          </EuiFlexGroup>
-        )}
+                </h6>
+              </EuiTitle>
+            </EuiText>
+          </EuiFlexItem>
+        </EuiFlexGroup>
       </EuiPopoverTitle>
     );
   };