+++ public/types.ts
@@ -16,6 +16,8 @@ import {
 } from '../../../src/plugins/visualizations/public';
 import { DataPublicPluginSetup, DataPublicPluginStart } from '../../../src/plugins/data/public';
 import { AppMountParameters, CoreStart } from '../../../src/core/public';
+import { AssistantClient } from './services/assistant_client';
+import { UiActionsSetup, UiActionsStart } from '../../../src/plugins/ui_actions/public';
 import { UsageCollectionSetup } from '../../../src/plugins/usage_collection/public';
 
 export interface RenderProps {
@@ -40,6 +42,7 @@ export interface AssistantPluginStartDependencies {
   visualizations: VisualizationsStart;
   embeddable: EmbeddableStart;
   dashboard: DashboardStart;
+  uiActions: UiActionsStart;
 }
 
 export interface AssistantPluginSetupDependencies {
@@ -47,6 +50,7 @@ export interface AssistantPluginSetupDependencies {
   visualizations: VisualizationsSetup;
   embeddable: EmbeddableSetup;
   dataSourceManagement?: DataSourceManagementPluginSetup;
+  uiActions: UiActionsSetup;
   usageCollection?: UsageCollectionSetup;
 }
 
@@ -63,12 +67,14 @@ export interface AssistantSetup {
    */
   nextEnabled: () => boolean;
   assistantActions: Omit<AssistantActions, 'executeAction'>;
+  assistantTriggers: { AI_ASSISTANT_QUERY_EDITOR_TRIGGER: string };
   registerIncontextInsight: IncontextInsightRegistry['register'];
   renderIncontextInsight: (component: React.ReactNode) => React.ReactNode;
 }
 
 export interface AssistantStart {
   dataSource: DataSourceServiceContract;
+  assistantClient: AssistantClient;
 }
 
 export type StartServices = CoreStart &