+++ server/routes/summary_routes.ts
@@ -8,12 +8,17 @@ import { IRouter } from '../../../../src/core/server';
 import { SUMMARY_ASSISTANT_API } from '../../common/constants/llm';
 import { getOpenSearchClientTransport } from '../utils/get_opensearch_client_transport';
 import { getAgent, searchAgentByName } from './get_agent';
-import { ML_COMMONS_BASE_API } from '../utils/constants';
-import { InsightType, SummaryType } from '../types';
+import { AssistantServiceSetup } from '../services/assistant_service';
 
-const SUMMARY_AGENT_CONFIG_ID = 'summary';
+const SUMMARY_AGENT_CONFIG_ID = 'os_summary';
+const OS_INSIGHT_AGENT_CONFIG_ID = 'os_insight';
+let osInsightAgentId: string | undefined;
+let userInsightAgentId: string | undefined;
 
-export function registerSummaryAssistantRoutes(router: IRouter) {
+export function registerSummaryAssistantRoutes(
+  router: IRouter,
+  assistantService: AssistantServiceSetup
+) {
   router.post(
     {
       path: SUMMARY_ASSISTANT_API.SUMMARIZE,
@@ -34,39 +39,39 @@ export function registerSummaryAssistantRoutes(router: IRouter) {
         context,
         dataSourceId: req.query.dataSourceId,
       });
-      const agentId = await getAgent(SUMMARY_AGENT_CONFIG_ID, client);
-      const prompt = SummaryType.find((type) => type.id === req.body.type)?.prompt;
-      const response = await client.request({
-        method: 'POST',
-        path: `${ML_COMMONS_BASE_API}/agents/${agentId}/_execute`,
-        body: {
-          parameters: {
-            prompt,
-            context: req.body.context,
-            question: req.body.question,
-          },
-        },
+      const assistantClient = assistantService.getScopedClient(req, context);
+      const response = await assistantClient.executeAgentByName(SUMMARY_AGENT_CONFIG_ID, {
+        context: req.body.context,
+        question: req.body.question,
       });
       let summary;
-      let insightAgentId;
+      let insightAgentIdExists = false;
       try {
         if (req.body.insightType) {
           // We have separate agent for os_insight and user_insight. And for user_insight, we can
           // only get it by searching on name since it is not stored in agent config.
           if (req.body.insightType === 'os_insight') {
-            insightAgentId = await getAgent(req.body.insightType, client);
+            if (!osInsightAgentId) {
+              osInsightAgentId = await getAgent(OS_INSIGHT_AGENT_CONFIG_ID, client);
+            }
+            insightAgentIdExists = !!osInsightAgentId;
           } else if (req.body.insightType === 'user_insight') {
             if (req.body.type === 'alerts') {
-              insightAgentId = await searchAgentByName('KB_For_Alert_Insight', client);
+              if (!userInsightAgentId) {
+                userInsightAgentId = await searchAgentByName('KB_For_Alert_Insight', client);
+              }
             }
+            insightAgentIdExists = !!userInsightAgentId;
           }
         }
       } catch (e) {
-        console.log(`Cannot find insight agent for ${req.body.insightType}`);
+        context.assistant_plugin.logger.info(
+          `Cannot find insight agent for ${req.body.insightType}`
+        );
       }
       try {
         summary = response.body.inference_results[0].output[0].result;
-        return res.ok({ body: { summary, insightAgentId } });
+        return res.ok({ body: { summary, insightAgentIdExists } });
       } catch (e) {
         return res.internalError();
       }
@@ -77,7 +82,7 @@ export function registerSummaryAssistantRoutes(router: IRouter) {
       path: SUMMARY_ASSISTANT_API.INSIGHT,
       validate: {
         body: schema.object({
-          insightAgentId: schema.string(),
+          summaryType: schema.string(),
           insightType: schema.string(),
           summary: schema.string(),
           context: schema.string(),
@@ -93,18 +98,19 @@ export function registerSummaryAssistantRoutes(router: IRouter) {
         context,
         dataSourceId: req.query.dataSourceId,
       });
-      const prompt = InsightType.find((type) => type.id === req.body.insightType)?.prompt;
-      const response = await client.request({
-        method: 'POST',
-        path: `${ML_COMMONS_BASE_API}/agents/${req.body.insightAgentId}/_execute`,
-        body: {
-          parameters: {
-            text: prompt,
-            context: req.body.context,
-            summary: req.body.summary,
-            question: req.body.question,
-          },
-        },
+      const insightAgentId =
+        req.body.insightType === 'os_insight' ? osInsightAgentId : userInsightAgentId;
+      if (!insightAgentId) {
+        context.assistant_plugin.logger.info(
+          `Cannot find insight agent for ${req.body.insightType}`
+        );
+        return res.internalError();
+      }
+      const assistantClient = assistantService.getScopedClient(req, context);
+      const response = await assistantClient.executeAgent(insightAgentId, {
+        context: req.body.context,
+        summary: req.body.summary,
+        question: req.body.question,
       });
       try {
         let result = response.body.inference_results[0].output[0].result;
@@ -112,6 +118,9 @@ export function registerSummaryAssistantRoutes(router: IRouter) {
         return res.ok({ body: result });
       } catch (e) {
         return res.internalError();
+      } finally {
+        // Reset userInsightAgentId in case users update their insight agent.
+        userInsightAgentId = undefined;
       }
     })
   );